add jwt

+++ backend/controllers/blogs.js
@@ -1,6 +1,6 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response, next) => {
     try {
@@ -28,14 +28,13 @@ blogsRouter.get('/:id', async (request, response, next) => {
     }
 })
 
-blogsRouter.post('/', async (request, response, next) => {
+blogsRouter.post('/', userExtractor, async (request, response, next) => {
     try {
-        const { title, author, url, likes, userId } = request.body
-
-        const user = await User.findById(userId)
+        const { title, author, url, likes } = request.body
+        const user = request.user
 
         if (!user) {
-            return response.status(400).json({ error: 'userId missing or not valid' })
+            return response.status(401).json({ error: 'user not authenticated' })
         }
 
         const blog = new Blog({ title, author, url, likes, user: user._id })
@@ -45,39 +44,51 @@ blogsRouter.post('/', async (request, response, next) => {
         await user.save()
 
         response.status(201).json(savedBlog)
-    }
-    catch (exception) {
+    } catch (exception) {
         next(exception)
     }
 })
 
-blogsRouter.put('/:id', async (request, response, next) => {
+blogsRouter.put('/:id', userExtractor, async (request, response, next) => {
     try {
-        const blog = request.body
+        const blog = await Blog.findById(request.params.id)
+
+        if (!blog) {
+            return response.status(404).json({ error: 'blog not found' })
+        }
+
+        if (blog.user.toString() !== request.user.id.toString()) {
+            return response.status(403).json({ error: 'unauthorized' })
+        }
 
         const updatedBlog = await Blog.findByIdAndUpdate(
             request.params.id,
-            blog,
-            { returnDocument: 'after', runValidators: true, context: 'query' }
+            request.body,
+            { new: true, runValidators: true, context: 'query' }
         )
 
-        if (updatedBlog) {
-            response.json(updatedBlog)
-        } else {
-            response.status(404).end()
-        }
-    }
-    catch (exception) {
+        response.json(updatedBlog)
+    } catch (exception) {
         next(exception)
     }
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response, next) => {
     try {
+        const blog = await Blog.findById(request.params.id)
+
+        if (!blog) {
+            return response.status(404).json({ error: 'blog not found' })
+        }
+
+        if (blog.user.toString() !== request.user.id.toString()) {
+            return response.status(403).json({ error: 'unauthorized' })
+        }
+
         await Blog.findByIdAndDelete(request.params.id)
+
         response.status(204).end()
-    }
-    catch (exception) {
+    } catch (exception) {
         next(exception)
     }
 })
